pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["p79"]

[project]
name = "p79"
version = "0.1.0"
description = "Project 79: Cost-Aware Routing for Web Usage Agents with VisualWebArena + Qwen3-VL-4B"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision",
    "transformers>=4.37.0",
    "accelerate>=0.26.0",
    "bitsandbytes>=0.42.0",
    "qwen-vl-utils",
    "pillow",
    "numpy",
    "pyyaml",
    "pandas",
    "matplotlib",
    "scipy",
    "einops"
]

[project.scripts]
run_episode = "scripts.run_episode:main"
